+++ src/lib/validateCsv.ts
@@ -13,7 +13,7 @@ export interface Diagnostic {
 export type Diagnostics = Diagnostic[];
 
 /** Validate CSV data against layout definition */
-export async function validateData(
+export async function validateCsv(
   conn: duckdb.AsyncDuckDBConnection,
   headers: string[],
   layout: Layout,
